backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
scipy>=1.12.0
pydantic>=2.6.0
python-multipart>=0.0.9
google-genai>=1.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
